mapper implementation

+++ src/main/java/com/iws_manager/iws_manager_api/dtos/orderemployee/OrderEmployeeResponseDTO.java
@@ -15,7 +15,7 @@ public record OrderEmployeeResponseDTO(
     BigDecimal plannedhours,
     String qualificationkmui,
     String title,
-    Long version,
+    Integer version,
     
     EmployeeBasicDTO employee,
     OrderReferenceDTO order,
